added manu-sed-plotter in api + bug fixes

+++ main_plot-api.py
@@ -9,8 +9,9 @@
 #
 
 
-# Standard (important) packages
+# Standard packages
 import os
+import re
 
 # Custom packages
 import analyze_r3d_functions as a3d
@@ -70,25 +71,28 @@ app.layout = dbc.Container([
         options=[
             {'label':modelname, 'value':modelname}  for modelname in modelnames
         ],
-        placeholder='Chose a model'
+        placeholder='Chose a model',
+        value=0
     ),
 
     # Menu of folders inside chosen model
     dcc.Dropdown(
         id='phase-dropdown', 
         className='',
-        placeholder='Chose a phase'
+        placeholder='Chose a phase',
+        value=0
     ),
 
     html.Hr(),
     # Plot singular SED below here
     html.P('SEDs:'),
 
-    # Menu of SED-files
+    # Menu of Inclinations
     dcc.Dropdown(
-        id='sed-dropdown', 
+        id='incl-dropdown', 
         className='',
-        placeholder='Chose an SED'
+        placeholder='Chose an Inclination',
+        value='N/A'
     ),
 
     # SED plot
@@ -107,7 +111,8 @@ app.layout = dbc.Container([
     dcc.Dropdown(
         id='image-dropdown', 
         className='',
-        placeholder='Chose an image'
+        placeholder='Chose an image',
+        value=0
     ),
 
     # Image plot
@@ -140,7 +145,8 @@ app.layout = dbc.Container([
             {'label':'All phases (constant inclination)', 'value':'allphases'},
             {'label':'All inclinations (constant phase)', 'value':'allincls'}
         ],
-        labelStyle={'display': 'block'}
+        labelStyle={'display': 'block'},
+        value=0
     ),
 
     # Plot all chosen SEDs
@@ -148,7 +154,8 @@ app.layout = dbc.Container([
         id='sed-plot-all'
     ),
     html.P('SED-luminosity (Lsol):'),
-    html.Div(id='sed_luminosity_average'),
+    # TODO
+    #html.Div(id='sed_luminosity_average'),
 
 
 
@@ -186,53 +193,57 @@ app.layout = dbc.Container([
 )
 def create_phase_dict(modelname):
 
-    phases = [
-        phase for phase in os.listdir(path=f'{path}{modelname}/') if os.path.isdir(f'{path}{modelname}/{phase}') == True
-    ]
+    if modelname != 0 :
+        phases = [
+            phase for phase in os.listdir(path=f'{path}{modelname}/') if os.path.isdir(f'{path}{modelname}/{phase}') == True
+        ]
 
-    if len(phases) > 0:
-        phases.sort()
+        if len(phases) > 0:
+            phases.sort()
+        else:
+            phases = ['No phases found']
 
-    else:
-        phases = ['No phases found']
-
-    return [{'label':phase, 'value':phase} for phase in phases]
+        return [{'label':phase, 'value':phase} for phase in phases]
 
 
 # Given model and phase, add list of images and SEDs
 @app.callback(
     Output('image-dropdown', 'options'),
-    Output('sed-dropdown', 'options'),
+    Output('incl-dropdown', 'options'),
     Input('model-dropdown', 'value'),
     Input('phase-dropdown', 'value'),
 )
 def create_image_sed_dicts(modelname,phase):
 
-    # Extract file names in folder
-    filenames = os.listdir(path=f'{path}{modelname}/{phase}')
-    filenames.sort()
+    if modelname != 0 or phase != 0:
+
+        # Extract file names in folder
+        filenames = os.listdir(path=f'{path}{modelname}/{phase}')
+        filenames.sort()
 
 
-    # Extract image file names
-    images = [image for image in filenames if image[0:5] == 'image']
+        # Extract image file names
+        images = [image for image in filenames if image[0:5] == 'image']
 
-    # Check if there are any
-    if len(images) > 0:
-        images.sort()
-    else:
-        images = ['No image found']
+        # Check if there are any
+        if len(images) > 0:
+            images.sort()
+        else:
+            images = ['No image found']
 
+        # Extract inclinations of SEDs
+        incls = []
+        for filename in filenames:
+            if filename[:8] == 'spectrum':
+                incls.append(re.findall('spectrum_i.*.', filename)[0][10:-4])
+        
+        # Check if there are any SEDs
+        if len(incls) > 0:
+            incls.sort()
+        else:
+            incls = ['N/A']
 
-    # Extract SED file names
-    seds = [sed for sed in filenames if sed[0:8] == 'spectrum']
-
-    # Check if there are any SEDs
-    if len(seds) > 0:
-        seds.sort()
-    else:
-        seds = ['No SED found']
-
-    return [{'label':image, 'value':image} for image in images], [{'label':sed, 'value':sed} for sed in seds]
+        return [{'label':image, 'value':image} for image in images], [{'label':f'Inclination: {incl} deg', 'value':incl} for incl in incls]
 
 
 # Given model, phase and sed-file, plot this
@@ -241,43 +252,48 @@ def create_image_sed_dicts(modelname,phase):
     Output('sed_luminosity', 'children'),
     Input('model-dropdown', 'value'),
     Input('phase-dropdown', 'value'),
-    Input('sed-dropdown', 'value'),
+    Input('incl-dropdown', 'value'),
 )
-def plot_sed_one(modelname,phase,sed):
+def plot_sed_one(modelname,phase,incl):
 
-    # Extract luminosity of SED
-    lum = a3d.compute_sed_luminosity(
-        path=f'{path}{modelname}/{phase}/{sed}'
-    )/3.828e26
+    if modelname != 0 or phase != 0 or incl != 'N/A':
 
-    # Extract image
-    fig,ax,maxflux,maxwave = a3d.plot_sed(
-        path=f'{path}{modelname}/{phase}/{sed}'
-    )
-    plotly_fig = tls.mpl_to_plotly(fig)
+        # Extract luminosity of SED
+        lum = a3d.compute_sed_luminosity(
+            path=f'{path}{modelname}/{phase}/spectrum_i{incl}.out'
+        )/3.828e26
 
-    # Change some settings on the plotly plot
-    plotly_fig.update_layout(
-        title_font_size=18,
-        hoverlabel_font_size=18,
-        plot_bgcolor='white',
-        yaxis = dict(tickfont = dict(size=16)),
-        xaxis = dict(tickfont = dict(size=16))
-    )
-    plotly_fig.update_xaxes(
-        titlefont_size=16,
-        showgrid=True, gridwidth=1, gridcolor='lightgrey',
-        ticks="outside", tickwidth=2, ticklen=10,
-        showline=True, linewidth=2, linecolor='black', mirror=False
-    )
-    plotly_fig.update_yaxes(
-        titlefont_size=16,
-        showgrid=True, gridwidth=1, gridcolor='lightgrey',
-        ticks="outside", tickwidth=2, ticklen=10,
-        showline=True, linewidth=2, linecolor='black', mirror=False
-    )
+        # Extract image
+        fig,ax,maxflux,maxwave = a3d.plot_sed(
+            path=f'{path}{modelname}/{phase}/spectrum_i{incl}.out'
+        )
+        plotly_fig = tls.mpl_to_plotly(fig)
 
-    return plotly_fig, lum
+        # Change some settings on the plotly plot
+        plotly_fig.update_layout(
+            title_font_size=18,
+            hoverlabel_font_size=18,
+            plot_bgcolor='white',
+            yaxis = dict(tickfont = dict(size=16)),
+            xaxis = dict(tickfont = dict(size=16))
+        )
+        plotly_fig.update_xaxes(
+            titlefont_size=16,
+            showgrid=True, gridwidth=1, gridcolor='lightgrey',
+            ticks="outside", tickwidth=2, ticklen=10,
+            showline=True, linewidth=2, linecolor='black', mirror=False
+        )
+        plotly_fig.update_yaxes(
+            titlefont_size=16,
+            showgrid=True, gridwidth=1, gridcolor='lightgrey',
+            ticks="outside", tickwidth=2, ticklen=10,
+            showline=True, linewidth=2, linecolor='black', mirror=False
+        )
+
+        return plotly_fig, lum
+
+
+
 
 
 # Given model, phase and image-file, plot this
@@ -289,83 +305,111 @@ def plot_sed_one(modelname,phase,sed):
 )
 def plot_image_one(modelname,phase,image):
 
-    # If image already exists, remove it so that the image updates
-    if os.path.exists('temp.png') == True:
-        os.system('rm temp.png')
+    if modelname != 0 or phase != 0 or image != 0:
 
-    # Change image to list
-    image = [image]
+        # If image already exists, remove it so that the image updates
+        if os.path.exists('temp.png') == True:
+            os.system('rm temp.png')
 
-    # Load and save image as png
-    fig,ax,testflux = a3d.plot_images(
-        path = f'{path}{modelname}/{phase}/',
-        images = image,
-        distance = 1
-    )
-    plt.savefig(
-        'temp.png', dpi=120, bbox_inches='tight'
-    )
+        # Change image to list
+        image = [image]
 
-    # Change to base64-encoding that html.Img can read
-    tempbase64 = base64.b64encode(open('temp.png', 'rb').read()).decode('ascii')
-    impath = 'data:image/png;base64,{}'.format(tempbase64)
+        # Load and save image as png
+        fig,ax,testflux = a3d.plot_images(
+            path = f'{path}{modelname}/{phase}/',
+            images = image,
+            distance = 1
+        )
+        plt.savefig(
+            'temp.png', dpi=120, bbox_inches='tight'
+        )
 
-    return impath
+        # Change to base64-encoding that html.Img can read
+        tempbase64 = base64.b64encode(open('temp.png', 'rb').read()).decode('ascii')
+        impath = 'data:image/png;base64,{}'.format(tempbase64)
 
+        return impath
 
-
-# TODO finish this!
+"""
 # Given model, phase and all phases or all inclinations, plot all seds of choice
 @app.callback(
+    Output('sed-plot-all', 'figure'),
     Input('model-dropdown', 'value'),
     Input('phase-dropdown', 'value'),
-    Input('sed-picker', 'value')
+    Input('incl-dropdown', 'value'),
+    Input('sed-picker', 'value'),
 )
-def plot_sed_all(modelname,phase,choice):
+def plot_sed_all(modelname,phase,incl,choice):
 
 
+    if modelname != 0 or phase != 0 or incl != 'N/A' or choice != 0:
 
+
+    # pathlist and legendlist depends on choice.
+    pathlist = []
+    legendlist = []
+
+    # All phases, ie inclination is chosen earlier, take all phases with this
+    if choice == 'allphases':
+
+        legendlist = [
+            phase for phase in os.listdir(path=f'{path}{modelname}/') if os.path.isdir(f'{path}{modelname}/{phase}') == True
+        ]
+
+        for phase in legendlist:
+            pathlist.append(f'{path}{modelname}/{phase}/spectrum_i{incl}.out')
+
+    # Else, fill list with paths to chosen phsae and all incls there instead
+    if choice == 'allincls':
+        
+        filenames = os.listdir(f'{path}{modelname}/{phase}/')
+
+        for filename in filenames:
+            if filename[:8] == 'spectrum':
+                legendlist.append(re.findall('spectrum_i.*.', filename)[0][10:-4])
+
+        for incl in legendlist:
+            pathlist.append(f'{path}{modelname}/{phase}/spectrum_i{incl}.out')
 
     # Create SED figure
     fig,ax = a3d.plot_sedsmany(
         pathlist=pathlist,
-        legendlist=legendlist,
         distance=1
     )
-    plotly_fig = tls.mpl_to_plotly(fig)
+    plotly_fig_all = tls.mpl_to_plotly(fig)
 
     # TODO compute average luminosity of chosen SEDs
 
 
 
     # Change some settings on the plotly plot
-    plotly_fig.update_layout(
+    plotly_fig_all.update_layout(
         title_font_size=18,
         hoverlabel_font_size=18,
         plot_bgcolor='white',
         yaxis = dict(tickfont = dict(size=16)),
         xaxis = dict(tickfont = dict(size=16))
     )
-    plotly_fig.update_xaxes(
+    plotly_fig_all.update_xaxes(
         titlefont_size=16,
         showgrid=True, gridwidth=1, gridcolor='lightgrey',
         ticks="outside", tickwidth=2, ticklen=10,
         showline=True, linewidth=2, linecolor='black', mirror=False
     )
-    plotly_fig.update_yaxes(
+    plotly_fig_all.update_yaxes(
         titlefont_size=16,
         showgrid=True, gridwidth=1, gridcolor='lightgrey',
         ticks="outside", tickwidth=2, ticklen=10,
         showline=True, linewidth=2, linecolor='black', mirror=False
     )
 
-    return plotly_fig
+    return plotly_fig_all
 
-
-
+"""
 
 
 
 if __name__ == "__main__":
     app.run_server(debug=True)
+    #app.run_server(port=8050)
